+++ src/components/Homepage/index.js
@@ -1,5 +1,4 @@
 export { default as Clock } from './Clock';
-export { default as HomeScheduleList } from './HomeScheduleList';
+export { default as ScheduleTimeLine } from './ScheduleTimeLine';
 export { default as HomeTodoList } from './HomeTodoList';
-export { default as ScheduleBox } from './ScheduleBox';
-export { default as ScheduleDetail } from './ScheduleDetail';
+export { default as TodoItem } from './TodoItem';
